+++ src/materials/blinn.cpp
@@ -22,9 +22,9 @@ void blinn::sampleF(const float& u0, const float& u1, const vec3& wo, vec3& wi, 
     wh.normalize();
 
     if(wh.y > 0){
-        wi = 2.f * fabs(dot(wo, wh)) * wh - wo;
+        wi = 2.f * fabsf(dot(wo, wh)) * wh - wo;
 
-        pd = ((exp + 2.f) * powf(cosThetaH, exp)) / (4.f * TWOPI * fabs(dot(wo, wh)));
+        pd = ((exp + 2.f) * powf(cosThetaH, exp)) / (4.f * TWOPI * fabsf(dot(wo, wh)));
     }else{
         pd = 0.f;
     }
